requirements.txt
pandas==2.2.3
numpy==2.1.3
yfinance==0.2.50
mysql-connector-python==9.1.0
et-xmlfile==2.0.0 
openpyxl==3.1.5
